move CountryFlag selection in dedicated comp for reusability
