+++ components/Navbar/Navbar.module.scss
@@ -1,7 +1,7 @@
 @import "@/styles/theme.scss";
 
 .container {
-  @apply relative flex-1 h-16 p-4 bg-black items-center justify-between text-white;
+  @apply flex-1 h-16 p-4 bg-black items-center justify-between text-white;
 
   .items {
     @apply flex-1 space-x-6 text-sm ml-8 h-6 items-center;
@@ -20,13 +20,8 @@
   }
 }
 
-.frame {
-  @apply relative;
-  @apply items-center justify-center rounded h-8 w-8 cursor-pointer;
-
-  span {
-    @apply font-bold text-sm;
-  }
+.countryFlag {
+  @apply h-8 w-8 rounded;
 }
 
 .floater {
@@ -50,10 +45,6 @@
       @apply flex flex-row items-center space-x-2 hover:bg-black pl-3 pr-6 py-2;
     }
   }
-}
-
-.countryFlag {
-  @apply h-8 w-8 rounded;
 }
 
 .dropdownMenu {
