+++ components/Navbar/Settings.tsx
@@ -4,11 +4,7 @@ import { useTranslation } from "react-i18next";
 import Link from "next/link";
 import userMock from "@/tests/__mocks__/user"; // MOCK
 import useHover from "@react-hook/hover";
-import { langs } from "@/locales/i18n";
-import CountryFlag from "@/components/CountryFlag";
-import useSelector from "@/hooks/useSelector";
-import useDispatch from "@/hooks/useDispatch";
-import { setLang } from "@/state/users/actions";
+import { LangSettings } from "@/components/CountryFlag";
 import UserIcon from "@/components/UserIcon";
 import styles from "./Navbar.module.scss";
 
@@ -45,49 +41,6 @@ const UserLinks = () => {
     <FlexCol className={styles.floaterContent}>
       <Link href="account">{t("components.navbar.account")}</Link>
       <Link href="logout">{t("components.navbar.logout")}</Link>
-    </FlexCol>
-  );
-};
-
-const LangSettings = () => {
-  const currentLang = useSelector((state) => state.user.lang) as string;
-  const hoverableNode = React.useRef(null);
-  const isHovering = useHover(hoverableNode);
-
-  return (
-    <div ref={hoverableNode} className={styles.frame} data-testid="lang-flag">
-      <CountryFlag lang={currentLang} className={styles.countryFlag} />
-
-      {isHovering && (
-        <div className={styles.floater}>
-          <LangOptions />
-        </div>
-      )}
-    </div>
-  );
-};
-
-const LangOptions = () => {
-  const dispatch = useDispatch();
-  const { t, i18n } = useTranslation();
-
-  const changeLanguage = (lang: string) => {
-    i18n.changeLanguage(lang);
-    dispatch(setLang(lang));
-  };
-
-  return (
-    <FlexCol className={styles.floaterContent}>
-      {Object.entries(langs).map(([langKey, langValue]) => (
-        <button
-          key={langKey}
-          type="button"
-          onClick={() => changeLanguage(langKey)}
-        >
-          <CountryFlag lang={langKey} width={15} height={15} />
-          <span>{t(langValue)}</span>
-        </button>
-      ))}
     </FlexCol>
   );
 };
